Fixes.
Rename `ScraperConfig` -> `Config`.
`Config` fields made al public.
Created Warnings Channel.
Created a way to be able for the pkg user to handle the warnings and errors.
Moved some of the errors sent on `errors` channel to warnings (this might need some changing to what is sent through which channel).